requirements.txt
docopt
pyenchant
nltk
spacy
-e .